A função genPlane respeita as divisões

+++ Generator/plane.hpp
@@ -1,2 +1,2 @@
 #include <math.h>
-void genPlane(float x, float z, float *v);
+void genPlane(float x, int z, float *v);
